requirements.txt
fastapi
uvicorn
gunicorn
supabase
ortools
requests
python-dotenv
notion-client
